API: add `encoding` arg, return unicode string from `graph_from_dot_file`
because dot files are text, and should be represented as `unicode`
in python 2, and `str` in python 3.
Explicit is better than implicit,
so the caller should provide the encoding.

In general, there is no way to recognize an encoding.
There are ways to guess an encoding (for example,
using the package `chardet`), but those are heuristics
that can fail, thus not dependable for production code.

Use the `io.open` function, which is the same in both
python 2 and 3, and serves as the backported builtin
`open` of python 3 [2]. Unlike the builtin `open` of
python 2, `io.open` distinguishes between unicode
(as `unicode` in python 2, and `str` in python 3)
and bytes (as `str` in python 2, and `bytes` in python 3).

[1] https://docs.python.org/2/library/io.html
[2] https://mail.python.org/pipermail/python-list/2015-March/687124.html