TST: sdist and then install with `pip` on travis
to test that all necessary files have been
included in `MANIFEST.in`, and so avoid errors like:

b4b44a9046785d446136896d16f84357c11a1fb4

This error was more complicated than what it looks like.
It escaped both locally and on Travis,
because of installing from source.
Locally with:

`python setup.py install`

and on Travis with:

`pip install .`

Instead, first create a distribution archive,
and then install from it, as done in this commit.

In more detail, files mentioned in `MANIFEST.in`
but missing result only in warnings during sdist.
In the error referenced by the commit hash above,
the missing file was a renamed README.

Upon install (locally or on Travis),
the missing file may still go unnoticed,
because it may have been mentioned only in `MANIFEST.in`.

If the file is mentioned also elsewhere,
in our case it was in `data_files` within `setup.py`,
then `pip install dist/*.tar.gz` fails with:

"error: can't copy 'README.md': doesn't exist or not a regular file"