MAI: rm `data_files` from `setup.py`
because the files listed: `LICENSE` and `README.md` are
included in `MANIFEST.in` and need not be installed to
site-packages.

Also, as explained in the message of commit:

2f172917a5e39db2c335a386e0c42dac06f82851

this mention of (non-data) files raises hard failures,
even though they are not essential to a functional installation.

Their absence should have been detected,
but earlier, during sdist (which only warns).
For example, using the package `check-manifest`.