TST: replace CRLF with LF in 3 test files
These files were previously committed with CRLF line endings.
Whatever line endings they have in the working directory,
the current `.gitattributes` instruct `git` to commit LF line
endings. So the directory is in an "unstable" state,
in that it regards itself as modified in absence of changes to
the working directory.

Close #163.