Prevent TypeError in handling of DOT parser error
When parsing DOT strings or files, pydot's parser calls the third-party
library PyParsing. If PyParsing meets a problem, it raises a
`ParseException`, which pydot then uses to report some details.

Currently, during the handling of that exception, a `TypeError` occurs.
It is caused by an attempt to concatenate a string with the exception
using the `+` (addition/plus) operator, which is [not allowed][1]. The
problem was introduced in 2016 by commit [b4a38103][2].

This current commit now removes that concatenation and goes back to
printing three separate lines. The end result is the same as Lance
Hepler's [original 2013 solution][3] and [the example code in the
PyParsing documentation][4].

I tested that behavior is now as expected (printing parser error
details without `TypeError`) under Python 2.7.16 with PyParsing 2.4.2
and Python 3.7.3 with PyParsing 2.4.2 and 3.0.0a2 ([486b1fd2e][5]).
Unit tests all pass (same three combinations, when run together with
the test suite fix proposed in PR 211).

This [closes #176][6] ("Error in error handling hides the real error
message"), reported by Shish, and resolves the `TypeError` part of
[issue #171][7] ("Exception on parsing dot file"), reported by Michael
Goerz.

[1]: https://docs.python.org/3/reference/expressions.html#binary-arithmetic-operations
[2]: https://github.com/pydot/pydot/commit/b4a38103a553cbc7f0fa3fdd530e8e1fcfdba7f9#diff-baef597193866f900a2726a8a4667b12R563-R566
[3]: https://github.com/nlhepler/pydot/commit/687eb7e18c3d0cc890557e8874b13f7f1b802667#diff-baef597193866f900a2726a8a4667b12R551-R553
[4]: https://github.com/pyparsing/pyparsing/commit/2c6f881943ecf0fbf02354623b51cc0566325c75#diff-7ab8c6f0d66d170e10b60b9f65334e18R694-R697
[5]: https://github.com/pyparsing/pyparsing/tree/486b1fd2ef7e98966665f915bc59856996ffb5b0
[6]: https://github.com/pydot/pydot/issues/176
[7]: https://github.com/pydot/pydot/issues/171