Improve traceback error message on GraphViz errors
For reporting GraphViz errors, improve the "AssertionError" message
in the traceback by including program name and arguments. This is the
same as the first line of the error message that is printed to stdout,
but not followed by the (potentially long) Graphviz stdout and stderr
output.

This is a short term improvement, since an AssertionError is not
ideal. See also https://github.com/pydot/pydot/issues/203