Dyn methods: Use setattr
Not seeing any customizations that necessitate the use of the magic
method `__setattr__()` here, so using `setattr()` should be equivalent.
[[1]] [[2]]

Timeit is reporting 20 to 80us reductions in object instantation times,
all 20% drops relative to pydot 1.4.1.

No other changes in external behavior are expected.

[1]: https://stackoverflow.com/questions/14756289/setattrobject-name-value-vs-object-setattr-name-value
[2]: https://stackoverflow.com/questions/7559170/whats-the-difference-between-setattr-and-object-setattr

(This commit is part of a series of changes in how the `get_*`,
`set_*`, `create_*` and `write_*` methods are dynamically created.)