Add test for simple/simplified graphs
This test tests the output of `Graph.to_string()` for graphs with
multiple (parallel) edges and combinations of `simplify` set to `True`
or `False` and the graph being directed or undirected.

This test was originally created for issue [pydot/pydot#92][1]
("Simplify throws NameError, and, as implemented, doesn't work"). It
can also help catch regressions when doing future work on
`Graph.to_string()`, `Edge.__eq__()` or `Edge.__hash__()`.

[1]: https://github.com/pydot/pydot/issues/92