Stop using `str` as a function parameter name
`dot_parser.py` currently has 6 functions that still use `str` as a
parameter name. This will cause conflicts if inside the function body
a reference to Python's built-in [`str` type][1] needs to be made.

Specifically, as part of the clean up of old Python 2 code, the next
commit will introduce direct usage of the built-in `str` in two of
these functions. The concurrent use of `str` as a function parameter
name would then lead to name collisions and errors such as:

      File "dot_parser.py", line 349, in push_edge_stmt
        elif isinstance(toks[2][0], str):
    TypeError: isinstance() arg 2 must be a type or tuple of types

Commit 842173cab already renamed the parameter for one particular
function from `str` to `s` in 2008. This commit will now do the same
for the remaining functions.

Further notes:
- None of the affected functions actually use the parameter (`str`, now
  `s`) in their function bodies.
- The functions themselves are passed to `pyparsing.setParseAction()`
  as callback functions. The docstring of that method also uses the
  parameter name `s` in the example callback signature. However, the
  actual parameter name is irrelevant for pyparsing (2.4.7), as it uses
  positional arguments when actually calling the callback.
- The affected functions contain 3 format strings that use `{s}` as a
  replacement field name. Other than perhaps being slightly disturbing
  to the human reader, these do not actually collide with the function
  parameters named `s`, so leaving them as-is for now. A future change
  to f-strings would be a better way to resolve the confusion.

[1]: https://docs.python.org/3/library/stdtypes.html#str