Drop Python 2 and 3.4: Clean up PY2-specific code
Just a very basic cleanup, made with some help from 2to3.

Includes the removal of intermediairy variable `str_type`, which held
the string type for either Python 2 or 3, replacing it with the string
type for Python 3 `str`. The previous commit prepared for this by
removing the conflicting uses of the name `str` as function parameter
names.