Remove working directory check from test suite
This commit removes a check from the test suite that was introduced by
commit 50bd7308 ([pydot/pydot#164][164]). The check raised an error in
case the current working directory was the top of the source tree. The
purpose was to prevent testing the pydot sources from the source tree
instead of the installed pydot.

The risk of that inadvertently happening is considerably lower now that
the sources were moved to subdirectory `src/`. Also note that [Python
_running a script_ does not by default search the working directory][1]
anyway.

The removal of the check makes the `--no-check` option redundant.
However, we cannot remove it yet because that would cause an
`unrecognized arguments` error in existing scripts. `argparse` does not
support [deprecating CLI arguments][2] yet. So, instead we print a
deprecation warning. The remaining `parse_args()` and `import argparse`
can be removed in a later major release of pydot.

[1]: https://chrisyeh96.github.io/2017/08/08/definitive-guide-python-imports.html
[2]: https://bugs.python.org/issue39467
[164]: https://github.com/pydot/pydot/issues/164