Fix needs_quotes() for values containing commas
The regular expressions underlying needs_quotes() were treating the
comma as a valid character for numeric values (presumably implying
European-style comma decimals). But in graphviz, the comma is an
attribute separator, so values containing it should always be quoted,
even if it is the current locale's decimal separator.

Replace regular expressions like `[0-9,]*` with `[0-9\.]*`, to ensure
string values containing commas are always quoted, and periods in string
values don't force quoting.