Fix parenting behavior for duplicate-name nodes (#365)
* Fix parenting of duplicate nodes

If add_node() is called with a Node that has the same name as an
existing node, it's append()ed to the list of nodes hashed under
that name in the obj_dict. When that's done, set_parent_graph()
isn't called on the node (perhaps under the assumption that it
may already have a different parent graph). We should check that,
and still call set_parent_graph() if its parent is None.

* Add tests for node parenting