Fix pickling of Dot objects
The instance variables of the class also need to be pickled, along
with the obj_dict, so that they aren't reset to defaults when
unpickled. (Or worse, left completely unset.)

For backwards compatibility, a directly-pickled `obj_dict` will be
recognized and restored as such, with the other variables being
reset to defaults.