Eliminate parser's top_graphs global
Turns out it isn't needed, as the list can simply be created
locally in `push_top_graph_stmt` and returned from there.

Note that this WON'T make the parser thread-safe, as pyparsing
itself isn't thread-safe. But it'll remove a data race in our
parser module, when handling the output of the parser.

See #401