Vendor in Python 3.12 tempfile.py, for < 3.10
Python 3.8 and 3.9 are missing the `ignore_cleanup_errors` option
to `tempfile.TemporaryDirectory()`, which is needed to ensure that
spurious Windows tempfile cleanup errors don't lead to crashes.

Include a copy of the Python 3.12 `tempfile.py` (slightly modified
to be compatible with Python 3.8 `types` and `shutil.rmtree`), and
use that instead of the standard library module to ensure that we
can use `ignore_cleanup_errors` in all Python versions.