Add __eq__ and __ne__ methods to FrozenDict
This complements the new `__hash__` method with sorting removed; one
of the Python requirements for hashability is that "objects which
compare equal have the same hash value".

So these two should not be equal, unlike for `dict.__eq__`:
`FrozenDict({'a': 1, 'b': 2}) != FrozenDict({'b': 2, 'a': 1})`

(We need FrozenDicts to be hashable because they're used in tuples
that become the hash keys for a Graph's `obj_dict['edges']`. So, the
order of their items should matter.)