Remove Common.__get_attribute__
Since the dawn of time (the first pydot commit, 17 years ago), the
Common class has included a `__get_attribute__` method. It's always
been completely untested (0 coverage), and we don't use it for
anything.

It was likely _intended_ to provide (Python) attribute-based access
to the attributes of the graph members, but doesn't actually do that
since the dunder method for retrieving attributes is `__getattribute__`
(no space between the words).

It's probably a good thing that it doesn't work, because it contains
some weird and suspicious code for dealing with defaults that I don't
think we'd actually want to implement that way. Let's just chuck it.