Implement .resolve() method for attributes
Where .get(name) will only look up the value for the 'name' attribute
if it's set on the object for which it's called, .resolve(name) will
look for that attribute, and if it's not found, walk up the graph
hierarchy scanning for a relevant default value. If one is found, that
value is returned instead.