Simplify get_top_graph_type with default (#437)
To make use of get_top_graph_type simpler for callers, add an
optional argument "default", with a default value of "graph".
This value will be returned if the element has no identifiable
top graph type (e.g. an Edge not yet added to a Graph).

"graph" was chosen as the default (even though the _actual_
default type for graphs is "digraph") because it generally
triggers more expansive processing in code that relies on it.
The tools to look up edges, for example, will use the endpoints
in both orderings, when in an undirected "graph" context.

This also makes the typing easier, as `get_top_graph_type` now
returns `str` in all cases.