Fix numeric-value regular expression
The `re_all_numeric` check used to allow any number of decimal
points in the value, which isn't correct for numerics; at most one
can be allowed. Rename to `re_numeric`, and use a more complex
RE that allows a maximum of one decimal point, but fails on strings
like "1.4.0".

Fixes #438