Edge constructor: Fix construction from tuple
One of the pickling tests contained a call of the form
`pydot.Graph.add_edge(pydot.Edge(("D", "E")))`
which we never noticed didn't actually work as expected. The
resulting graph threw a traceback when its `.to_string()` was
called, because the edge added was from `("D", "E") -- ""`,
which wasn't stringified properly by `.to_string()`. The Edge
constructor will now accept a list or tuple as the first arg.