Port API tests to pytest (#453)
* tests: Port TestGraphApi suite to pytest

- Remove unittest class TestGraphApi from test_pydot.py
- Create new test_api.py, containing tests from TestGraphApi suite
  ported to pytest
- Add graph_directed fixture to conftest.py

* Edge constructor: Fix construction from tuple

One of the pickling tests contained a call of the form
`pydot.Graph.add_edge(pydot.Edge(("D", "E")))`
which we never noticed didn't actually work as expected. The
resulting graph threw a traceback when its `.to_string()` was
called, because the edge added was from `("D", "E") -- ""`,
which wasn't stringified properly by `.to_string()`. The Edge
constructor will now accept a list or tuple as the first arg.

* tests: Fix test_node_parenting lookup

* tests: Moving logging test to separate file

* tests: Remove unused imports from test_pydot

* Dumb things down for Python 3.8