dot_parser: Move parser definition into class
Instead of a global variable `graphparser` that's set by calling
a function the first time it's referenced, set up the parser
definition in a class `GraphParser`, so that it can be accessed
as `GraphParser.parser`.

This has the advantage that all of its _component_ parts (like
the definition of `subgraph_stmt`, or `attr_list` can also
be accessed as `GraphParser.subgraph_stmt` or
`GraphParser.attr_list`, etc. (Much better for debugging/testing.)

In addition, to improve dependency relationships:
- Only import pydot.core into dot_parser, not `pydot`
- Change all `pydot.Foo` references to `pydot.core.Foo`
- Use concrete `pydot.core.Foo`s in type annotations
  (instead of strings)