dot_parser: Define parser elements in a class (#464)
* core: Import dot_parser on demand

* dot_parser: Move parser definition into class

Instead of a global variable `graphparser` that's set by calling
a function the first time it's referenced, set up the parser
definition in a class `GraphParser`, so that it can be accessed
as `GraphParser.parser`.

This has the advantage that all of its _component_ parts (like
the definition of `subgraph_stmt`, or `attr_list` can also
be accessed as `GraphParser.subgraph_stmt` or
`GraphParser.attr_list`, etc. (Much better for debugging/testing.)

In addition, to improve dependency relationships:
- Only import pydot.core into dot_parser, not `pydot`
- Change all `pydot.Foo` references to `pydot.core.Foo`
- Use concrete `pydot.core.Foo`s in type annotations
  (instead of strings)

* Remove unused args from parse actions

* parser: Remove unnecessary type-ignore from HTML.__init__

It annoyed mypy in strict mode.

* dot_parser: Define 'graphparser' at module level

Strictly for compatibility, in case anyone is using it directly.
They probably shouldn't be, but this way we won't break their code.

* Move low-end MyPy check to Python 3.9