Add __future__.annotations imports (#478)
* Add __future__.annotations imports

Conversation in #473 led to the discovery that in Python 3.7-3.9,
forward-referencing annotations were only supported in the presence
of this import. Add it to make type-annotating code easier in the
future.

This then required adjusting _many_ of our types, with Ruff suddenly
insisting on `foo | None` syntax instead of `Optional[foo]`, and on
use of real container types in annotations rather than the `typing`
import representations (e.g. `list[str]` not `typing.List[str]`).
I think the appearance of `__future__.annotations` imports pushed it
to upgrade the implicit minimum version it uses for syntax checks.

* dot_parser: Import typing.Any exclusively

As a result of the previous commit, the _only_ symbol still used from
the `typing` namespace is `typing.Any`. So, instead of importing
all of `typing`, just import that one type.